+++ tests/rpc/test_rpc_client.py
@@ -1,13 +1,11 @@
 from __future__ import annotations
 
-import json
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, AsyncIterator, Dict, Optional
 
 import pytest
 
-from chia.cmds.cmds_util import get_any_service_client
 from chia.rpc.rpc_client import ResponseFailureError, RpcClient
 from chia.util.ints import uint16
 from tests.util.misc import Marks, RecordingWebServer, datacases
@@ -19,6 +17,15 @@ class InvalidCreateCase:
     root_path: Optional[Path] = None
     net_config: Optional[Dict[str, Any]] = None
     marks: Marks = ()
+
+
+@pytest.fixture(name="rpc_client")
+async def rpc_client_fixture(recording_web_server: RecordingWebServer) -> AsyncIterator[RpcClient]:
+    async with RpcClient.create_as_context(
+        self_hostname=recording_web_server.web_server.hostname,
+        port=recording_web_server.web_server.listen_port,
+    ) as rpc_client:
+        yield rpc_client
 
 
 @datacases(
@@ -51,21 +58,6 @@ async def test_rpc_client_works_without_ssl(recording_web_server: RecordingWebSe
     assert result == expected_result
 
 
-# TODO: think about where these tests actually belong
-
-sample_traceback = "this\nthat"
-sample_traceback_json = json.dumps(sample_traceback)
-
-
-@pytest.fixture(name="rpc_client")
-async def rpc_client_fixture(recording_web_server: RecordingWebServer) -> AsyncIterator[RpcClient]:
-    async with RpcClient.create_as_context(
-        self_hostname=recording_web_server.web_server.hostname,
-        port=recording_web_server.web_server.listen_port,
-    ) as rpc_client:
-        yield rpc_client
-
-
 @pytest.mark.anyio
 async def test_rpc_client_send_request(
     rpc_client: RpcClient,
@@ -87,70 +79,3 @@ async def test_failure_exception(
         await rpc_client.fetch(path="/table", request_json={"response": expected_response})
 
     assert exception_info.value.response == expected_response
-
-
-@pytest.mark.anyio
-async def test_failure_output_no_traceback(
-    root_path_populated_with_config: Path,
-    recording_web_server: RecordingWebServer,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    expected_response = {"success": False, "magic": "statue"}
-
-    async with get_any_service_client(
-        client_type=RpcClient,
-        rpc_port=recording_web_server.web_server.listen_port,
-        root_path=root_path_populated_with_config,
-        use_ssl=False,
-    ) as (client, _):
-        await client.fetch(path="/table", request_json={"response": expected_response})
-
-    out, err = capsys.readouterr()
-
-    assert "ResponseFailureError" not in out
-    assert "Traceback:" not in out
-    assert json.dumps(expected_response) in out
-
-
-@pytest.mark.anyio
-async def test_failure_output_with_traceback(
-    root_path_populated_with_config: Path,
-    recording_web_server: RecordingWebServer,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    expected_response = {"success": False, "traceback": sample_traceback}
-
-    async with get_any_service_client(
-        client_type=RpcClient,
-        rpc_port=recording_web_server.web_server.listen_port,
-        root_path=root_path_populated_with_config,
-        use_ssl=False,
-    ) as (client, _):
-        await client.fetch(path="/table", request_json={"response": expected_response})
-
-    out, err = capsys.readouterr()
-    assert sample_traceback_json not in out
-    assert sample_traceback in out
-
-
-@pytest.mark.anyio
-async def test_failure_output_no_consumption(
-    root_path_populated_with_config: Path,
-    recording_web_server: RecordingWebServer,
-    capsys: pytest.CaptureFixture[str],
-) -> None:
-    expected_response = {"success": False, "magic": "xylophone"}
-
-    with pytest.raises(ResponseFailureError) as exception_info:
-        async with get_any_service_client(
-            client_type=RpcClient,
-            rpc_port=recording_web_server.web_server.listen_port,
-            root_path=root_path_populated_with_config,
-            use_ssl=False,
-            consume_errors=False,
-        ) as (client, _):
-            await client.fetch(path="/table", request_json={"response": expected_response})
-
-    assert exception_info.value.response == expected_response
-
-    assert capsys.readouterr() == ("", "")
